% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_dados_cov_ind.R
\name{baixar_dados_cov_ind}
\alias{baixar_dados_cov_ind}
\title{Baixar e consolidar séries temporais econômicas}
\usage{
baixar_dados_cov_ind(start_date = "2011-01-01", end_date = Sys.Date())
}
\arguments{
\item{start_date}{Data inicial no formato "YYYY-MM-DD"}

\item{end_date}{Data final no formato "YYYY-MM-DD"}
}
\value{
Uma lista com os objetos `responses_ts` e `covariates_ts`
}
\description{
Esta função baixa e organiza dados de diversas fontes públicas.
}
